consolidate filenames, array types, input filetypes and rois
