+++ mbo_utilities/graphics/__init__.py
@@ -1,6 +1,9 @@
+# force import to ensure glfw is initialized before fastplotlib
+import glfw
+from rendercanvas.glfw import GlfwRenderCanvas, loop
+
 from .run_gui import run_gui
 from .imgui import PreviewDataWidget
-
 
 __all__ = [
     "PreviewDataWidget",
